+++ streaming/src/main/scala/org/apache/spark/streaming/scheduler/JobSet.scala
@@ -17,9 +17,9 @@
 
 package org.apache.spark.streaming.scheduler
 
-import scala.collection.mutable.{ArrayBuffer, HashSet}
+import scala.collection.mutable.HashSet
+
 import org.apache.spark.streaming.Time
-import org.apache.spark.streaming.receiver.ReceivedBlockInfo
 
 /** Class representing a set of Jobs
   * belong to the same batch.
