+++ streaming/src/main/scala/org/apache/spark/streaming/scheduler/ReceivedBlockInfo.scala
@@ -1,7 +1,8 @@
-package org.apache.spark.streaming.receiver
+package org.apache.spark.streaming.scheduler
 
 import org.apache.spark.storage.StreamBlockId
 import org.apache.spark.streaming.util.WriteAheadLogFileSegment
+import org.apache.spark.annotation.DeveloperApi
 
 /** Information about blocks received by the receiver */
 private[streaming] case class ReceivedBlockInfo(
